Moved all classes to io.pactfolw.example.sirenconsumer
